Headers no longer require a child anchor element
In the accordion control.

http://jqueryui.com/upgrade-guide/1.9/#headers-no-longer-require-a-child-anchor-element

BREAKING CHANGE:
As a result of the above change the Hyperlink object has been removed.
If you wish to still include an a.href you can by setting it as normal
in "Title" or the "Panel".

See test "Accordion_Can_Set_ID_On_Panel" for an example.