Position: offset option deprecitated
The "offset" option in the position utility class has been dropped, as
the "my" option now supports offsets.  Full details at:

http://jqueryui.com/upgrade-guide/1.9/#deprecated-offset-option-merged-into-my-and-at

Note this has an impact on the Fluqi API for the AutoComplete and Dialog
objects which no longer have an offset option.  Please upgrade your code
to move any offset declarations into the "my" property as illustrated in
the above change log.